openai streaming frontend
